refactor: rename bookmark to shortcut
